refactor(web): Phase 2 — lift projects into 3-tier (Infra → Service → API)
Second domain migrated under `docs/web-architecture.md`'s API → Services →
Infra layout (issue #313). Projects was chosen as the next step because it
introduces the shared `GitClient` infra alongside the per-domain project
state, exercising the cross-domain composition the architecture doc
describes — the `projects.remove` route still composes cronjob teardown at
the API layer rather than burying it inside the service.

- Infra: `server/infra/git/git-client.ts` — extract every shell-out to
  `git` from `lib/git.ts` into a `GitClient` class plus a stable
  functional surface (`execGit`, `execGh`, `gitCmd`, `listWorktrees`,
  `getRepoInfo`, `parseGitRemoteUrl`, `DETACHED_BRANCH_PREFIX`) so
  services / routers / sync loops can depend on a typed, single-purpose
  adapter. Documented infra-tier rules in the file header.
- Infra: `server/infra/db/queries/projects.ts` — extract the projects +
  worktrees DB CRUD (`loadAll`, `saveAll`, `setHasOrigin`) into a
  `ProjectQueries` class, plus the `ProjectState` / `WorktreeState` types
  and the `reconcileKindForProject` helper. The whole-tree `saveAll`
  rewrite + focused `setHasOrigin` UPDATE pair mirrors the prior
  `lib/state` semantics — see `setHasOrigin` JSDoc for the race motivation
  (issue #458).
- Service: `server/services/project-service.ts` — own all business logic
  that used to be inline in `trpc/router.ts`: `list` (kind reconciliation
  + `git worktree list` enrichment + status / label join), `add` (kind
  detection, label validation, default branch resolution, worktree seed),
  `promoteToGit` (plain → git escape hatch from #427), `remove`, `reorder`,
  `updateLabel`, `checkPath`, `gitInit`, `worktreesDir`. Module-level
  `projectService` singleton consumed by the API tier.
- API: `server/api/projects/router.ts` — thin tRPC sub-router that
  validates input with Zod and delegates to the service. The `remove`
  route is the one place this router composes more than a single service
  call — project-scoped cronjob teardown (`stopJobsForKey` +
  `deleteCronjobFile`) stays at the API layer per the architecture doc.
- API merge point: `server/api/router.ts` now composes the migrated
  `projectsRouter` alongside `settingsRouter` via `t.mergeRouters(...)`.
  Invariant comment expanded so future phases keep the merge surface
  unambiguous.
- Legacy router shrunk: removed the inline `projectsRouter` declaration
  from `trpc/router.ts` and dropped now-unused `loadCurrentStatuses` /
  `ProjectKind` / `reconcileKindForProject` / `basename` imports. The
  `projects.*` wire surface is unchanged — clients keep hitting the same
  tRPC paths through the merged router.
- Back-compat: `lib/state.ts` re-exports `ProjectState`, `WorktreeState`,
  `ProjectKind`, and `reconcileKindForProject` from the new infra module,
  and the `loadState` / `saveState` / `setProjectHasOrigin` wrappers now
  delegate to `ProjectQueries`. `lib/git.ts` re-exports the entire
  `GitClient` surface from the new infra module. Existing callers
  (sync-state, branch-status-poller, cronjob-scheduler, workspace.ts,
  github-graphql, the legacy tRPC router) keep compiling unchanged.

Closes #313
